+++ cfgov/cfgov/urls.py
@@ -6,17 +6,34 @@ from django.views.generic.base import TemplateView
 from sheerlike.views.generic import SheerTemplateView
 from sheerlike.feeds import SheerlikeFeed
 
-from v1.views import LeadershipCalendarPDFView, EventICSView, unshare, renderDirectoryPDF
+from v1.views import LeadershipCalendarPDFView, unshare, renderDirectoryPDF, \
+    change_password, password_reset_confirm, cfpb_login
 
 from wagtail.wagtailadmin import urls as wagtailadmin_urls
 from wagtail.wagtaildocs import urls as wagtaildocs_urls
 from wagtail.wagtailcore import urls as wagtail_urls
 from django.views.generic import RedirectView
 
+from wagtail.wagtailadmin.forms import PasswordResetForm
+from wagtail.wagtailadmin.views import account
 
 urlpatterns = [
+    url(r'^django-admin/login', cfpb_login, name='wagtailadmin_login'),
+    url(r'^django-admin/password_change', change_password, name='wagtailadmin_account_change_password'),
     url(r'^django-admin/', include(admin.site.urls)),
     url(r'^admin/pages/(\d+)/unshare/$', unshare, name='unshare'),
+
+    # - Overridded Wagtail Password views - #
+    url(r'^admin/login/$', cfpb_login, name='wagtailadmin_login'),
+    url(r'^admin/password_reset/', include([
+        url(
+            r'^confirm/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$',
+            password_reset_confirm, name='wagtailadmin_password_reset_confirm',
+        )
+    ])),
+    url(r'^admin/account/change_password/$', change_password, name='wagtailadmin_account_change_password'),
+    # ----------------x-------------------- #
+
     url(r'^admin/', include(wagtailadmin_urls)),
     url(r'^documents/', include(wagtaildocs_urls)),
     # TODO: Enable search route when search is available.